stats_streamlit/requirements.txt
streamlit==1.37.1
psycopg2-binary==2.9.9
pandas==2.2.2
passlib
bcrypt==3.2.2
plotly==5.22.0
openpyxl==3.1.2
